kernelci.configs: convert to kernelci.config.{build,test}
* convert kernelci.configs into a new kernelci.config package with
  kernelci.config.build and kernelci.config.test

* update kci_build and lava-v2-jobs-from-api.py accordingly

* update unit tests accordingly
