+++ tests/configs/builds-empty-arch.yaml
