+++ tests/configs/builds-minimal.yaml
